lab5/lab5_py/requirements.txt
flask
requests
hazelcast-python-client
psutil
python-consul